+++ packages/react-native/src/posthog-rn.ts
@@ -26,52 +26,56 @@ import {
 } from './types'
 import { withReactNativeNavigation } from './frameworks/wix-navigation'
 import { OptionalReactNativeSessionReplay } from './optional/OptionalSessionReplay'
-import { ErrorTracking } from './error-tracking'
+import { ErrorTracking, ErrorTrackingOptions } from './error-tracking'
 
 export { PostHogPersistedProperty }
 
-export type PostHogOptions = PostHogCoreOptions &
-  PostHogRNOptions & {
-    /** Allows you to provide the storage type. By default 'file'.
-     * 'file' will try to load the best available storage, the provided 'customStorage', 'customAsyncStorage' or in-memory storage.
-     */
-    persistence?: 'memory' | 'file'
-    /** Allows you to provide your own implementation of the common information about your App or a function to modify the default App properties generated */
-    customAppProperties?:
-      | PostHogCustomAppProperties
-      | ((properties: PostHogCustomAppProperties) => PostHogCustomAppProperties)
-    /** Allows you to provide a custom asynchronous storage such as async-storage, expo-file-system or a synchronous storage such as mmkv.
-     * If not provided, PostHog will attempt to use the best available storage via optional peer dependencies (async-storage, expo-file-system).
-     * If `persistence` is set to 'memory', this option will be ignored.
-     */
-    customStorage?: PostHogCustomStorage
+export interface PostHogOptions extends PostHogCoreOptions {
+  /** Allows you to provide the storage type. By default 'file'.
+   * 'file' will try to load the best available storage, the provided 'customStorage', 'customAsyncStorage' or in-memory storage.
+   */
+  persistence?: 'memory' | 'file'
+  /** Allows you to provide your own implementation of the common information about your App or a function to modify the default App properties generated */
+  customAppProperties?:
+    | PostHogCustomAppProperties
+    | ((properties: PostHogCustomAppProperties) => PostHogCustomAppProperties)
+  /** Allows you to provide a custom asynchronous storage such as async-storage, expo-file-system or a synchronous storage such as mmkv.
+   * If not provided, PostHog will attempt to use the best available storage via optional peer dependencies (async-storage, expo-file-system).
+   * If `persistence` is set to 'memory', this option will be ignored.
+   */
+  customStorage?: PostHogCustomStorage
 
-    /** Captures app lifecycle events such as Application Installed, Application Updated, Application Opened, Application Became Active and Application Backgrounded.
-     * By default is false.
-     * Application Installed and Application Updated events are not supported with persistence set to 'memory'.
-     */
-    captureAppLifecycleEvents?: boolean
+  /** Captures app lifecycle events such as Application Installed, Application Updated, Application Opened, Application Became Active and Application Backgrounded.
+   * By default is false.
+   * Application Installed and Application Updated events are not supported with persistence set to 'memory'.
+   */
+  captureAppLifecycleEvents?: boolean
 
-    /**
-     * Enable Recording of Session Replays for Android and iOS
-     * Requires Record user sessions to be enabled in the PostHog Project Settings
-     * Defaults to false
-     */
-    enableSessionReplay?: boolean
+  /**
+   * Enable Recording of Session Replays for Android and iOS
+   * Requires Record user sessions to be enabled in the PostHog Project Settings
+   * Defaults to false
+   */
+  enableSessionReplay?: boolean
 
-    /**
-     * Configuration for Session Replay
-     */
-    sessionReplayConfig?: PostHogSessionReplayConfig
+  /**
+   * Configuration for Session Replay
+   */
+  sessionReplayConfig?: PostHogSessionReplayConfig
 
-    /**
-     * If enabled, the session id ($session_id) will be persisted across app restarts.
-     * This is an option for back compatibility, so your current data isn't skewed with the new version of the SDK.
-     * If this is false, the session id will be always reset on app restart.
-     * Defaults to false
-     */
-    enablePersistSessionIdAcrossRestart?: boolean
-  }
+  /**
+   * If enabled, the session id ($session_id) will be persisted across app restarts.
+   * This is an option for back compatibility, so your current data isn't skewed with the new version of the SDK.
+   * If this is false, the session id will be always reset on app restart.
+   * Defaults to false
+   */
+  enablePersistSessionIdAcrossRestart?: boolean
+
+  /**
+   * Error Tracking Configuration
+   */
+  errorTracking?: ErrorTrackingOptions
+}
 
 export class PostHog extends PostHogCore {
   private _persistence: PostHogOptions['persistence']
@@ -123,7 +127,7 @@ export class PostHog extends PostHogCore {
     this._persistence = options?.persistence ?? 'file'
     this._disableSurveys = options?.disableSurveys ?? false
     this._disableRemoteConfig = options?.disableRemoteConfig ?? false
-    this._errorTracking = new ErrorTracking(this)
+    this._errorTracking = new ErrorTracking(this, options?.errorTracking, this._logger)
 
     // Either build the app properties from the existing ones
     this._appProperties =
@@ -178,12 +182,12 @@ export class PostHog extends PostHogCore {
       if (this._disableRemoteConfig === false) {
         this.reloadRemoteConfigAsync()
       } else {
-        this.logMsgIfDebug(() => console.info('PostHog Debug', `Remote config is disabled.`))
+        this._logger.info('Remote config is disabled.')
         if (options?.preloadFeatureFlags !== false) {
-          this.logMsgIfDebug(() => console.info('PostHog Debug', `Feature flags will be preloaded from Flags API.`))
+          this._logger.info('Feature flags will be preloaded from Flags API.')
           this.reloadFeatureFlags()
         } else {
-          this.logMsgIfDebug(() => console.info('PostHog Debug', `preloadFeatureFlags is disabled.`))
+          this._logger.info('preloadFeatureFlags is disabled.')
         }
       }
 
@@ -234,6 +238,7 @@ export class PostHog extends PostHogCore {
   getLibraryVersion(): string {
     return version
   }
+
   getCustomUserAgent(): string {
     if (Platform.OS === 'web') {
       return ''
@@ -718,21 +723,14 @@ export class PostHog extends PostHogCore {
       if (OptionalReactNativeSessionReplay) {
         try {
           this._resetSessionId(OptionalReactNativeSessionReplay, String(sessionId))
-          this.logMsgIfDebug(() =>
-            console.info('PostHog Debug', `sessionId rotated from ${this._currentSessionId} to ${sessionId}.`)
-          )
+          this._logger.info(`sessionId rotated from ${this._currentSessionId} to ${sessionId}.`)
         } catch (e) {
-          this.logMsgIfDebug(() => console.error('PostHog Debug', `Failed to rotate sessionId: ${e}.`))
+          this._logger.error(`Failed to rotate sessionId: ${e}.`)
         }
       }
       this._currentSessionId = sessionId
     } else {
-      this.logMsgIfDebug(() =>
-        console.log(
-          'PostHog Debug',
-          `sessionId not rotated, sessionId ${sessionId} and currentSessionId ${this._currentSessionId}.`
-        )
-      )
+      this._logger.info(`sessionId not rotated, sessionId ${sessionId} and currentSessionId ${this._currentSessionId}.`)
     }
 
     return sessionId
@@ -743,9 +741,9 @@ export class PostHog extends PostHogCore {
     if (this._isEnableSessionReplay() && OptionalReactNativeSessionReplay) {
       try {
         OptionalReactNativeSessionReplay.endSession()
-        this.logMsgIfDebug(() => console.info('PostHog Debug', `Session replay ended.`))
+        this._logger.info(`Session replay ended.`)
       } catch (e) {
-        this.logMsgIfDebug(() => console.error('PostHog Debug', `Session replay failed to end: ${e}.`))
+        this._logger.error(`Session replay failed to end: ${e}.`)
       }
     }
   }
@@ -793,14 +791,9 @@ export class PostHog extends PostHogCore {
         distinctId = distinctId || previousDistinctId
         const anonymousId = this.getAnonymousId()
         OptionalReactNativeSessionReplay.identify(String(distinctId), String(anonymousId))
-        this.logMsgIfDebug(() =>
-          console.info(
-            'PostHog Debug',
-            `Session replay identified with distinctId ${distinctId} and anonymousId ${anonymousId}.`
-          )
-        )
+        this._logger.info(`Session replay identified with distinctId ${distinctId} and anonymousId ${anonymousId}.`)
       } catch (e) {
-        this.logMsgIfDebug(() => console.error('PostHog Debug', `Session replay failed to identify: ${e}.`))
+        this._logger.error(`Session replay failed to identify: ${e}.`)
       }
     }
   }
@@ -853,7 +846,7 @@ export class PostHog extends PostHogCore {
 
   public async getSurveys(): Promise<SurveyResponse['surveys']> {
     if (this._disableSurveys === true) {
-      this.logMsgIfDebug(() => console.log('PostHog Debug', 'Loading surveys is disabled.'))
+      this._logger.info('Loading surveys is disabled.')
       this.setPersistedProperty<SurveyResponse['surveys']>(PostHogPersistedProperty.Surveys, null)
       return []
     }
@@ -861,10 +854,10 @@ export class PostHog extends PostHogCore {
     const surveys = this.getPersistedProperty<SurveyResponse['surveys']>(PostHogPersistedProperty.Surveys)
 
     if (surveys && surveys.length > 0) {
-      this.logMsgIfDebug(() => console.log('PostHog Debug', 'Surveys fetched from storage: ', JSON.stringify(surveys)))
+      this._logger.info('Surveys fetched from storage: ', JSON.stringify(surveys))
       return surveys
     } else {
-      this.logMsgIfDebug(() => console.log('PostHog Debug', 'No surveys found in storage'))
+      this._logger.info('No surveys found in storage')
     }
 
     if (this._disableRemoteConfig === true) {
@@ -882,7 +875,7 @@ export class PostHog extends PostHogCore {
   private async startSessionReplay(options?: PostHogOptions): Promise<void> {
     this._enableSessionReplay = options?.enableSessionReplay
     if (!this._isEnableSessionReplay()) {
-      this.logMsgIfDebug(() => console.info('PostHog Debug', 'Session replay is not enabled.'))
+      this._logger.info('Session replay is not enabled.')
       return
     }
 
@@ -919,9 +912,7 @@ export class PostHog extends PostHogCore {
       throttleDelayMs,
     }
 
-    this.logMsgIfDebug(() =>
-      console.log('PostHog Debug', `Session replay SDK config: ${JSON.stringify(sdkReplayConfig)}`)
-    )
+    this._logger.info(`Session replay SDK config: ${JSON.stringify(sdkReplayConfig)}`)
 
     // if Flags API has not returned yet, we will start session replay with default config.
     const sessionReplay = this.getPersistedProperty(PostHogPersistedProperty.SessionReplay) ?? {}
@@ -929,18 +920,10 @@ export class PostHog extends PostHogCore {
     const cachedFeatureFlags = (featureFlags as { [key: string]: FeatureFlagValue }) ?? {}
     const cachedSessionReplayConfig = (sessionReplay as { [key: string]: JsonType }) ?? {}
 
-    this.logMsgIfDebug(() =>
-      console.log(
-        'PostHog Debug',
-        `Session replay feature flags from flags cached config: ${JSON.stringify(cachedFeatureFlags)}`
-      )
-    )
+    this._logger.info('Session replay feature flags from flags cached config:', JSON.stringify(cachedFeatureFlags))
 
-    this.logMsgIfDebug(() =>
-      console.log(
-        'PostHog Debug',
-        `Session replay session recording from flags cached config: ${JSON.stringify(cachedSessionReplayConfig)}`
-      )
+    this._logger.info(
+      `Session replay session recording from flags cached config: ${JSON.stringify(cachedSessionReplayConfig)}`
     )
 
     let recordingActive = true
@@ -962,30 +945,21 @@ export class PostHog extends PostHogCore {
         recordingActive = false
       }
 
-      this.logMsgIfDebug(() =>
-        console.log('PostHog Debug', `Session replay '${linkedFlag}' linked flag value: '${value}'`)
-      )
+      this._logger.info(`Session replay '${linkedFlag}' linked flag value: '${value}'`)
     } else if (linkedFlag && typeof linkedFlag === 'object') {
       const flag = linkedFlag['flag'] as string | undefined
       const variant = linkedFlag['variant'] as string | undefined
       if (flag && variant) {
         const value = cachedFeatureFlags[flag]
         recordingActive = value === variant
-        this.logMsgIfDebug(() =>
-          console.log('PostHog Debug', `Session replay '${flag}' linked flag variant '${variant}' and value '${value}'`)
-        )
+        this._logger.info(`Session replay '${flag}' linked flag variant '${variant}' and value '${value}'`)
       } else {
         // disable recording if the flag does not exist/quota limited
-        this.logMsgIfDebug(() =>
-          console.log(
-            'PostHog Debug',
-            `Session replay '${flag}' linked flag variant: '${variant}' does not exist/quota limited.`
-          )
-        )
+        this._logger.info(`Session replay '${flag}' linked flag variant: '${variant}' does not exist/quota limited.`)
         recordingActive = false
       }
     } else {
-      this.logMsgIfDebug(() => console.log('PostHog Debug', `Session replay has no cached linkedFlag.`))
+      this._logger.info(`Session replay has no cached linkedFlag.`)
     }
 
     if (recordingActive) {
@@ -993,7 +967,7 @@ export class PostHog extends PostHogCore {
         const sessionId = this.getSessionId()
 
         if (sessionId.length === 0) {
-          this.logMsgIfDebug(() => console.warn('PostHog Debug', 'Session replay enabled but no sessionId found.'))
+          this._logger.warn(`Session replay enabled but no sessionId found.`)
           return
         }
 
@@ -1007,9 +981,7 @@ export class PostHog extends PostHogCore {
           flushAt: this.flushAt,
         }
 
-        this.logMsgIfDebug(() =>
-          console.log('PostHog Debug', `Session replay sdk options: ${JSON.stringify(sdkOptions)}`)
-        )
+        this._logger.info(`Session replay sdk options: ${JSON.stringify(sdkOptions)}`)
 
         try {
           if (!(await OptionalReactNativeSessionReplay.isEnabled())) {
@@ -1019,25 +991,21 @@ export class PostHog extends PostHogCore {
               sdkReplayConfig,
               cachedSessionReplayConfig
             )
-            this.logMsgIfDebug(() =>
-              console.info('PostHog Debug', `Session replay started with sessionId ${sessionId}.`)
-            )
+            this._logger.info(`Session replay started with sessionId ${sessionId}.`)
           } else {
             // if somehow the SDK is already enabled with a different sessionId, we reset it
             this._resetSessionId(OptionalReactNativeSessionReplay, String(sessionId))
-            this.logMsgIfDebug(() =>
-              console.log('PostHog Debug', `Session replay already started with sessionId ${sessionId}.`)
-            )
+            this._logger.info(`Session replay already started with sessionId ${sessionId}.`)
           }
           this._currentSessionId = sessionId
         } catch (e) {
-          this.logMsgIfDebug(() => console.error('PostHog Debug', `Session replay failed to start: ${e}.`))
+          this._logger.error(`Session replay failed to start: ${e}.`)
         }
       } else {
-        this.logMsgIfDebug(() => console.warn('PostHog Debug', 'Session replay enabled but not installed.'))
+        this._logger.warn('Session replay enabled but not installed.')
       }
     } else {
-      this.logMsgIfDebug(() => console.info('PostHog Debug', 'Session replay disabled.'))
+      this._logger.info('Session replay disabled.')
     }
   }
 
@@ -1056,12 +1024,10 @@ export class PostHog extends PostHogCore {
         | undefined
 
       if (!appBuild || !appVersion) {
-        this.logMsgIfDebug(() =>
-          console.warn(
-            'PostHog could not track installation/update/open, as the build and version were not set. ' +
-              'This can happen if some dependencies are not installed correctly, or if you have provided' +
-              'customAppProperties but not included $app_build or $app_version.'
-          )
+        this._logger.warn(
+          'PostHog could not track installation/update/open, as the build and version were not set. ' +
+            'This can happen if some dependencies are not installed correctly, or if you have provided' +
+            'customAppProperties but not included $app_build or $app_version.'
         )
       }
       if (appBuild) {
@@ -1079,10 +1045,8 @@ export class PostHog extends PostHogCore {
         }
       }
     } else {
-      this.logMsgIfDebug(() =>
-        console.warn(
-          'PostHog was initialised with persistence set to "memory", capturing native app events (Application Installed and Application Updated) is not supported.'
-        )
+      this._logger.warn(
+        'PostHog was initialised with persistence set to "memory", capturing native app events (Application Installed and Application Updated) is not supported.'
       )
     }